refactored test_qtik and SwapTx
